Redefined project goals
I was focusing too much on the responsive navigation and not what I really wanted for this project, which was learn how to build the memory game.

Responsive navigation will be it's own project now.  I left only the title on the index page and will follow along with Ania Kubow's tutorial